literal block replaced with code blocks
Using code blocks to render OPAL output seems to be better than literal
blocks. The same is true for excerpts of the OPAL input file.

Due to a bug in the asciidoc2pdf converter script the last line of a file
must be terminated with a newline. To ensure this a 'EOF' comment has been
added to all files.